refactor(scripts): move the preview-links resolver into @nimbus-ds/scripts
.scripts/storybook-preview-links.ts duplicated logic nimbus-patterns had to
hand-port verbatim into its own copy (nimbus-patterns#198) — real file
duplication the moment a second repo needed the same deep-link resolution.
Moved the resolver, comment builder and CLI into
packages/core/scripts/src/storybookPreviewLinks, parameterized by a
PreviewLinksConfig (componentRootPattern, optional stylesComponentPattern)
so each repo's own component-tree layout is the only thing left local. This
repo's .scripts/storybook-preview-links.ts is now an 8-line wrapper calling
runPreviewLinksCli with its own patterns.

Bumps @nimbus-ds/scripts to 1.9.0 (minor — new exports). Once this publishes,
nimbus-patterns can depend on it and delete its own hand-ported copy instead
of maintaining it in parallel; noted as a follow-up, not done here — bumping
that dependency before the version exists on npm would break `yarn install`
there.

Also fixes two review findings caught while at it: `gate` in
preview-storybook.yml had no explicit `permissions:` block (CodeRabbit,
CWE-732 on both PRs), and the "no link per component" comment text implied a
single cause when the story being absent from index.json is another.
