+++ .scripts/storybook-preview-links.ts
@@ -1,415 +1,16 @@
-import fs from "fs";
-import path from "path";
+import { runPreviewLinksCli } from "@nimbus-ds/scripts/src";
 
 /**
- * Renders the body of the "Storybook preview is ready" pull request comment,
- * deep-linking every component the pull request touches.
- *
- * Story ids are read from the `index.json` emitted by `storybook build` instead
- * of being guessed from the `title` of a `*.stories.tsx` file: Storybook's own
- * sanitizer collapses dots, slashes and repeated separators (`Atomic/Input/
- * Input.Password` becomes `atomic-input-input-password`), so any hand-rolled
- * slug eventually produces a dead link.
+ * Thin CLI wrapper: the actual resolution logic lives in `@nimbus-ds/scripts`
+ * (packages/core/scripts/src/storybookPreviewLinks), shared with
+ * nimbus-patterns' own wrapper of the same name. This file only supplies
+ * this repo's component-tree layout.
  *
  * Usage (reads the changed file list from a file, or from stdin):
  *   STORYBOOK_URL=https://host/index.html yarn preview:links changed-files.txt
  */
-
-const DEFAULT_INDEX_PATH = ".build-storybook/index.json";
-const MAX_LINKS = 8;
-const MAX_MATCHED_FILES = 10;
-const PREVIEW_TRIGGERS: PreviewTrigger[] = [
-  "build-inputs",
-  "label",
-  "unavailable-diff",
-];
-
-/** Every published component lives in its own directory under these two roots. */
-const COMPONENT_ROOT_PATTERN =
-  /^packages\/react\/src\/(?:atomic|composite)\/[^/]+/;
-
-/** Style definitions live in their own package, keyed by the component name. */
-const STYLES_COMPONENT_PATTERN =
-  /^packages\/core\/styles\/src\/packages\/(?:atomic|composite)\/([^/]+)/;
-
-/** Generated by `yarn build:docs`, never a reason to point at a component. */
-const IGNORED_FILE_PATTERN = /\.docs\.json$/;
-
-/** Lets the workflow find its own comment again without matching on prose. */
-export const COMMENT_MARKER = "<!-- nimbus-storybook-preview -->";
-
-export interface StorybookIndexEntry {
-  id: string;
-  title: string;
-  name: string;
-  importPath: string;
-  type: string;
-}
-
-export interface StorybookIndex {
-  entries: Record<string, StorybookIndexEntry>;
-}
-
-export interface StoryTarget {
-  title: string;
-  docsId: string | null;
-  storyId: string | null;
-}
-
-const DOCS_ENTRY_TYPE = "docs";
-const STORY_ENTRY_TYPE = "story";
-
-/**
- * A component with no docs page is linked through one of its stories, and which
- * one must not depend on the order Storybook happens to emit — `ProgressBar`
- * exports eleven. These are the baseline names the repo's story convention asks
- * for; anything else keeps the index order, which is stable in practice but
- * nothing this file can guarantee.
- */
-const PRIMARY_STORY_NAMES = ["Default", "Basic", "Playground"];
-
-interface StoryCandidate {
-  id: string;
-  name: string;
-}
-
-interface StoryTargetDraft {
-  title: string;
-  docsId: string | null;
-  stories: StoryCandidate[];
-}
-
-const primaryStoryRank = (name: string): number => {
-  const rank = PRIMARY_STORY_NAMES.indexOf(name);
-
-  return rank === -1 ? PRIMARY_STORY_NAMES.length : rank;
-};
-
-/** Array#sort is stable, so equal ranks keep the order Storybook emitted. */
-const pickStoryId = (stories: StoryCandidate[]): string | null =>
-  [...stories].sort(
-    (a, b) => primaryStoryRank(a.name) - primaryStoryRank(b.name)
-  )[0]?.id ?? null;
-
-const normalizePath = (importPath: string): string =>
-  importPath.replace(/^\.\//, "");
-
-const directoryOf = (filePath: string): string => path.posix.dirname(filePath);
-
-const isSelfOrDescendant = (ancestor: string, directory: string): boolean =>
-  directory === ancestor || directory.startsWith(`${ancestor}/`);
-
-const byDepthThenName = (a: string, b: string): number =>
-  a.split("/").length - b.split("/").length || a.localeCompare(b);
-
-/**
- * `index.json` is generated, but a truncated or hand-edited one can hold
- * anything, and this module's contract is to fall back to the preview root
- * rather than throw. Checked as a set instead of one guard per field, because
- * every one of these is either interpolated into the comment or compared:
- * `importPath` reaches `.replace`, `title` reaches `.localeCompare`.
- */
-const isUsableEntry = (entry: StorybookIndexEntry): boolean =>
-  typeof entry?.importPath === "string" &&
-  typeof entry?.title === "string" &&
-  typeof entry?.id === "string";
-
-/**
- * Groups the flat Storybook index by the stories file each entry came from,
- * keeping the docs page and the story worth linking to.
- */
-const groupByStoriesFile = (
-  index: StorybookIndex
-): Map<string, StoryTarget> => {
-  const drafts = new Map<string, StoryTargetDraft>();
-
-  for (const entry of Object.values(index.entries ?? {})) {
-    if (!isUsableEntry(entry)) continue;
-
-    const storiesFile = normalizePath(entry.importPath);
-    const draft = drafts.get(storiesFile) ?? {
-      title: entry.title,
-      docsId: null,
-      stories: [],
-    };
-
-    if (entry.type === DOCS_ENTRY_TYPE && !draft.docsId) {
-      draft.docsId = entry.id;
-    }
-    if (entry.type === STORY_ENTRY_TYPE) {
-      draft.stories.push({ id: entry.id, name: entry.name });
-    }
-
-    drafts.set(storiesFile, draft);
-  }
-
-  return new Map(
-    [...drafts].map(([storiesFile, draft]) => [
-      storiesFile,
-      {
-        title: draft.title,
-        docsId: draft.docsId,
-        storyId: pickStoryId(draft.stories),
-      },
-    ])
-  );
-};
-
-/** `fileUploader`, `FileUploader` and `file-uploader` all name one component. */
-const normalizeComponentName = (name: string): string =>
-  name.toLowerCase().replace(/-/g, "");
-
-/**
- * Maps a change in `@nimbus-ds/styles` back to the component it styles. Those
- * files live outside the component tree, so they are matched by name.
- */
-const resolveStyledComponentRoot = (
-  changedFile: string,
-  storiesFiles: string[]
-): string | null => {
-  const styledComponent = STYLES_COMPONENT_PATTERN.exec(changedFile)?.[1];
-  if (!styledComponent) return null;
-
-  const wanted = normalizeComponentName(styledComponent);
-
-  return (
-    storiesFiles
-      .map((storiesFile) => COMPONENT_ROOT_PATTERN.exec(storiesFile)?.[0])
-      .find(
-        (componentRoot): componentRoot is string =>
-          Boolean(componentRoot) &&
-          normalizeComponentName(path.posix.basename(String(componentRoot))) ===
-            wanted
-      ) ?? null
-  );
-};
-
-/**
- * Finds the stories file that best describes a changed file: the one sitting in
- * the closest ancestor directory, so a change to `InputPassword` links to the
- * `Input.Password` page rather than to `Input`.
- */
-const resolveStoriesFile = (
-  changedFile: string,
-  storiesFiles: string[]
-): string | null => {
-  if (storiesFiles.includes(changedFile)) return changedFile;
-
-  const componentRoot =
-    COMPONENT_ROOT_PATTERN.exec(changedFile)?.[0] ??
-    resolveStyledComponentRoot(changedFile, storiesFiles);
-  if (!componentRoot) return null;
-
-  const candidates = storiesFiles.filter((storiesFile) =>
-    isSelfOrDescendant(componentRoot, directoryOf(storiesFile))
-  );
-  if (candidates.length === 0) return null;
-
-  const changedDirectory = directoryOf(changedFile);
-  const [nearest] = candidates
-    .filter((storiesFile) =>
-      isSelfOrDescendant(directoryOf(storiesFile), changedDirectory)
-    )
-    .sort((a, b) => directoryOf(b).length - directoryOf(a).length);
-
-  if (nearest) return nearest;
-
-  // Files that sit outside every stories directory — the package manifest, the
-  // changelog — still belong to the component, so fall back to its entry point.
-  return [...candidates].sort(byDepthThenName)[0];
-};
-
-export const resolveStoryTargets = (
-  changedFiles: string[],
-  index: StorybookIndex
-): StoryTarget[] => {
-  const targetsByStoriesFile = groupByStoriesFile(index);
-  const storiesFiles = [...targetsByStoriesFile.keys()];
-  const resolved = new Map<string, StoryTarget>();
-
-  for (const changedFile of changedFiles) {
-    if (IGNORED_FILE_PATTERN.test(changedFile)) continue;
-
-    const storiesFile = resolveStoriesFile(changedFile, storiesFiles);
-    if (!storiesFile) continue;
-
-    const target = targetsByStoriesFile.get(storiesFile);
-    if (target) resolved.set(storiesFile, target);
-  }
-
-  return [...resolved.values()].sort((a, b) => a.title.localeCompare(b.title));
-};
-
-export const previewUrl = (baseUrl: string, target: StoryTarget): string => {
-  const storyPath = target.docsId
-    ? `/docs/${target.docsId}`
-    : `/story/${target.storyId}`;
-  const separator = baseUrl.includes("?") ? "&" : "?";
-
-  return `${baseUrl}${separator}path=${storyPath}`;
-};
-
-/**
- * Why the workflow decided to build. The decision itself stays in the workflow,
- * which is the only place that can make it before dependencies are installed —
- * this module only reports it.
- */
-export type PreviewTrigger = "build-inputs" | "label" | "unavailable-diff";
-
-const TRIGGER_REASONS: Record<PreviewTrigger, string> = {
-  "build-inputs":
-    "this pull request changes files that feed the Storybook bundle",
-  label:
-    "the `storybook-preview` label forces a preview, whatever the diff contains",
-  "unavailable-diff":
-    "the list of changed files could not be read, so the preview was built rather than skipped",
-};
-
-export interface PreviewDecision {
-  trigger: PreviewTrigger;
-  /** The files that matched, evidence for a `build-inputs` trigger. */
-  matched: string[];
-}
-
-/**
- * Folded away behind a `<summary>`: the decision is worth being able to audit,
- * but nobody should have to read it to reach the link.
- */
-const buildDecisionDetails = (
-  decision: PreviewDecision,
-  linkableCount: number
-): string[] => {
-  const lines = [
-    "<details>",
-    "<summary>Why this preview exists, and why these links</summary>",
-    "",
-    `**Built because** ${TRIGGER_REASONS[decision.trigger]}.`,
-  ];
-
-  if (decision.trigger === "build-inputs" && decision.matched.length > 0) {
-    lines.push("");
-    for (const file of decision.matched.slice(0, MAX_MATCHED_FILES)) {
-      lines.push(`- \`${file}\``);
-    }
-    if (decision.matched.length > MAX_MATCHED_FILES) {
-      lines.push(`- …and ${decision.matched.length - MAX_MATCHED_FILES} more`);
-    }
-  }
-
-  lines.push("");
-  lines.push(
-    linkableCount > 0
-      ? "**The links above** come from the `index.json` of this very build, picking the stories file closest to each changed file — so a change under a sub-component lands on the sub-component's page, not on its parent's."
-      : "**No link per component** because nothing under `packages/react` or `@nimbus-ds/styles` changed, so there is no component page to point at. The root link is the whole preview."
-  );
-  lines.push("", "</details>");
-
-  return lines;
-};
-
-export const buildCommentBody = (
-  baseUrl: string,
-  targets: StoryTarget[],
-  decision: PreviewDecision | null = null
-): string => {
-  const lines = [COMMENT_MARKER, "🚀✨ Your Storybook preview is ready!", ""];
-
-  const linkable = targets.filter((target) => target.docsId ?? target.storyId);
-
-  if (linkable.length > 0) {
-    lines.push("Jump straight to what this pull request touches:", "");
-    for (const target of linkable.slice(0, MAX_LINKS)) {
-      lines.push(`- 🔗 [${target.title}](${previewUrl(baseUrl, target)})`);
-    }
-    if (linkable.length > MAX_LINKS) {
-      lines.push(`- …and ${linkable.length - MAX_LINKS} more`);
-    }
-    lines.push("");
-  }
-
-  lines.push(`🔗 [View Storybook](${baseUrl}) — full preview`, "");
-
-  if (decision) {
-    lines.push(...buildDecisionDetails(decision, linkable.length), "");
-  }
-
-  lines.push("Happy reviewing! 🎉");
-
-  return `${lines.join("\n")}\n`;
-};
-
-const readLines = (raw: string): string[] =>
-  raw
-    .split("\n")
-    .map((line) => line.trim())
-    .filter(Boolean);
-
-const readChangedFiles = (filePath: string | undefined): string[] =>
-  readLines(
-    filePath ? fs.readFileSync(filePath, "utf8") : fs.readFileSync(0, "utf8")
-  );
-
-const isPreviewTrigger = (value: string): value is PreviewTrigger =>
-  PREVIEW_TRIGGERS.includes(value as PreviewTrigger);
-
-/**
- * The workflow states its decision through PREVIEW_TRIGGER, and the files that
- * matched through PREVIEW_MATCHED_FILES. An unset or unrecognised trigger drops
- * the details block rather than inventing a reason for it.
- */
-const readDecision = (): PreviewDecision | null => {
-  const trigger = process.env.PREVIEW_TRIGGER;
-  if (!trigger || !isPreviewTrigger(trigger)) return null;
-
-  const matchedPath = process.env.PREVIEW_MATCHED_FILES;
-  const matched =
-    matchedPath && fs.existsSync(matchedPath)
-      ? readLines(fs.readFileSync(matchedPath, "utf8"))
-      : [];
-
-  return { trigger, matched };
-};
-
-/**
- * Never throws: the upload has already succeeded by the time this runs, so a
- * root-only link beats failing the step and leaving the reviewer no comment.
- */
-const readIndex = (indexPath: string): StorybookIndex | null => {
-  if (!fs.existsSync(indexPath)) {
-    console.error(
-      `\x1b[33m Storybook index not found at ${indexPath}, linking the preview root only \x1b[0m`
-    );
-    return null;
-  }
-
-  try {
-    return JSON.parse(fs.readFileSync(indexPath, "utf8")) as StorybookIndex;
-  } catch (error) {
-    console.error(
-      `\x1b[33m Storybook index at ${indexPath} is unreadable, linking the preview root only: ${
-        (error as Error).message
-      } \x1b[0m`
-    );
-    return null;
-  }
-};
-
-const main = (): void => {
-  const baseUrl = process.env.STORYBOOK_URL;
-  if (!baseUrl) {
-    console.error("\x1b[33m STORYBOOK_URL is required \x1b[0m");
-    process.exit(1);
-    return;
-  }
-
-  const index = readIndex(process.env.STORYBOOK_INDEX ?? DEFAULT_INDEX_PATH);
-  const targets = index
-    ? resolveStoryTargets(readChangedFiles(process.argv[2]), index)
-    : [];
-
-  process.stdout.write(buildCommentBody(baseUrl, targets, readDecision()));
-};
-
-if (require.main === module) {
-  main();
-}
+runPreviewLinksCli({
+  componentRootPattern: /^packages\/react\/src\/(?:atomic|composite)\/[^/]+/,
+  stylesComponentPattern:
+    /^packages\/core\/styles\/src\/packages\/(?:atomic|composite)\/([^/]+)/,
+});
