+++ .yarn/versions/15cb6440.yml
@@ -1,2 +1,6 @@
+releases:
+  "@nimbus-ds/scripts": minor
+
 declined:
   - nimbus-design-system
+  - nimbus-helper
