+++ packages/core/scripts/src/index.ts
@@ -3,3 +3,4 @@ export * from "./packageBuilder";
 export * from "./componentsBuilder";
 export * from "./packagePublisher";
 export * from "./common";
+export * from "./storybookPreviewLinks";
